# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel perfume api",
        "description": "L5 Swagger OpenApi ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/admin/factor-product/{id}": {
            "get": {
                "summary": "گرفتن یک زیر کالا ",
                "operationId": "12f2fa0d5a270d59ffac7ce378e6e728",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "access token",
                        "in": "query",
                        "description": "ایدی در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/PerfumeBasedFactor"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            },
            "put": {
                "summary": "تغییر دادن عضوی از فاکتور اضافه شده ( فقط وضعیت فروش و تعداد قابل تغییر است)(فقط سوپر ادمین و سازنده فاکتور توانایی پاک کردن آن را دارند)",
                "operationId": "0ae51debd1cc43cf14fbc08e291126ca",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ایدی  در یو ال ال",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stock": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "enum": [
                                            "true",
                                            "false"
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "is_active": "false",
                                    "stock": "-1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            },
            "delete": {
                "summary": "حذف عضوری از  فاکتور موجود (فقط سوپر ادمین و سازنده فاکتور توانایی پاک کردن آن را دارند)",
                "operationId": "36f9e74cfd2056a18807012360e88caf",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ایدی عضو فاکتور در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PerfumeBasedFactor": {
                "title": "گرفتن یک زیر کالا",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "string"
                    },
                    "sold": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "string"
                    },
                    "warranty": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "deleted": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}